+++ src/components/DpResettableInput/DpResettableInput.vue
@@ -8,7 +8,7 @@
       :required="required"
       :rounded="rounded"
       @blur="$emit('blur', currentValue)"
-      @input="$emit('input', currentValue)"
+      @input="onInput"
       @enter="$emit('enter', currentValue)"
       @focus="$emit('focus')"
       :pattern="pattern"
@@ -135,6 +135,11 @@ export default {
   },
 
   methods: {
+    onInput (val) {
+      this.currentValue = val
+      this.$emit('input', val)
+    },
+
     resetValue () {
       this.$emit('reset')
     }